requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudscraper==1.2.71
comm==0.2.2
cssselect==1.3.0
debugpy==1.8.15
decorator==5.2.1
duckduckgo_search==8.1.1
executing==2.2.0
fake-useragent==2.2.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google_search_results==2.4.2
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==11.1.1
Pygments==2.19.2
pyparsing==3.2.3
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2025.7.34
requests==2.32.4
requests-html==0.10.0
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.34.2
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
w3lib==2.3.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
